Add onion_test module.
This patch adds a new private module named onion_test. The purpose of
this module is to have auto-imported functions available for property,
unit and common test suites.

Currently it provides access to the various base{16,32,64} encoders and
decoders with the little difference that they accept iolists rather than
only binaries.